+++ src/ai-bundle/tests/DependencyInjection/AiBundleTest.php
@@ -46,7 +46,23 @@ class AiBundleTest extends TestCase
             ],
         ]);
 
+        $this->assertTrue($container->hasAlias('Symfony\AI\Agent\AgentInterface'));
         $this->assertTrue($container->hasAlias('Symfony\AI\Agent\AgentInterface $myAgentAgent'));
+    }
+
+    public function testAgentHasTag()
+    {
+        $container = $this->buildContainer([
+            'ai' => [
+                'agent' => [
+                    'my_agent' => [
+                        'model' => ['class' => 'Symfony\AI\Platform\Bridge\OpenAi\Gpt'],
+                    ],
+                ],
+            ],
+        ]);
+
+        $this->assertArrayHasKey('ai.agent.my_agent', $container->findTaggedServiceIds('ai.agent'));
     }
 
     #[TestWith([true], 'enabled')]
@@ -307,6 +323,12 @@ class AiBundleTest extends TestCase
                     ],
                     'ollama' => [
                         'host_url' => 'http://127.0.0.1:11434',
+                    ],
+                    'cerebras' => [
+                        'api_key' => 'cerebras_key_full',
+                    ],
+                    'voyage' => [
+                        'api_key' => 'voyage_key_full',
                     ],
                 ],
                 'agent' => [