+++ src/platform/src/Message/MessageBag.php
@@ -11,13 +11,15 @@
 
 namespace Symfony\AI\Platform\Message;
 
+use Symfony\AI\Platform\Metadata\MetadataAwareTrait;
+
 /**
- * @final
- *
  * @author Christopher Hertel <mail@christopher-hertel.de>
  */
-class MessageBag implements MessageBagInterface
+class MessageBag implements \Countable
 {
+    use MetadataAwareTrait;
+
     /**
      * @var list<MessageInterface>
      */
@@ -60,7 +62,7 @@ class MessageBag implements MessageBagInterface
         return $messages;
     }
 
-    public function merge(MessageBagInterface $messageBag): self
+    public function merge(self $messageBag): self
     {
         $messages = clone $this;
         $messages->messages = array_merge($messages->messages, $messageBag->getMessages());